recipes/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.3.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels-redis-4.2.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-silk==5.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-tailwind==3.8.0
flake8==7.1.1
gprof2dot==2024.6.6
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
isort==5.13.2
kombu==5.4.2
mccabe==0.7.0
numpy==2.2.1
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.2.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
